feat(mcp-server): migrate to MCP Python SDK 2.x
Refactor the MCP server into a weknora_mcp package with separate client,
tool schemas, dispatcher, server factory, and transport modules. Adopt the
SDK v2 low-level Server constructor callbacks and bump the dependency to
mcp>=2,<3 so fresh installs work on the current stable SDK line.
