+++ mcp-server/__init__.py
@@ -5,7 +5,7 @@ WeKnora MCP Server Package
 A Model Context Protocol server that provides access to the WeKnora knowledge management API.
 """
 
-__version__ = "1.0.1"
+__version__ = "1.1.0"
 __author__ = "WeKnora Team"
 __description__ = "WeKnora MCP Server - Model Context Protocol server for WeKnora API"
