+++ mcp-server/requirements.txt
@@ -1,4 +1,4 @@
-mcp>=1.28,<2
+mcp>=2.0,<3
 requests>=2.31.0
 starlette>=0.27.0
 uvicorn>=0.24.0
