+++ mcp-server/setup.py
@@ -24,7 +24,7 @@ def read_requirements():
             ]
     except FileNotFoundError:
         return [
-            "mcp>=1.28,<2",
+            "mcp>=2.0,<3",
             "requests>=2.31.0",
             "starlette>=0.27.0",
             "uvicorn>=0.24.0",
@@ -33,7 +33,7 @@ def read_requirements():
 
 setup(
     name="weknora-mcp",
-    version="1.0.1",
+    version="1.1.0",
     author="WeKnora Team",
     author_email="support@weknora.com",
     description="WeKnora MCP Server - Model Context Protocol server for WeKnora API",
@@ -41,6 +41,7 @@ setup(
     long_description_content_type="text/markdown",
     url="https://github.com/NannaOlympicBroadcast/WeKnoraMCP",
     py_modules=["weknora_mcp_server", "upload_paths", "main", "run_server", "run", "test_module"],
+    packages=["weknora_mcp", "weknora_mcp.tools"],
     classifiers=[
         "Development Status :: 4 - Beta",
         "Intended Audience :: Developers",
