+++ mcp-server/weknora_mcp/__init__.py
@@ -1,0 +1,18 @@
+"""WeKnora MCP server package."""
+
+from weknora_mcp.client import WeKnoraClient
+from weknora_mcp.config import SERVER_VERSION
+from weknora_mcp.server import create_client, create_server
+from weknora_mcp.transports import main, run, run_http, run_sse, run_stdio
+
+__all__ = [
+    "WeKnoraClient",
+    "SERVER_VERSION",
+    "create_client",
+    "create_server",
+    "main",
+    "run",
+    "run_http",
+    "run_sse",
+    "run_stdio",
+]
