+++ mcp-server/weknora_mcp/tools/__init__.py
@@ -1,0 +1,6 @@
+"""MCP tool schemas and dispatch logic."""
+
+from weknora_mcp.tools.dispatcher import ToolDispatcher
+from weknora_mcp.tools.schemas import TOOLS, build_tools
+
+__all__ = ["ToolDispatcher", "TOOLS", "build_tools"]
